+++ workspace/src/main/java/com/doctorTreat/app/doctorMypage/DoctorMypageFrontController.java
@@ -1,4 +1,4 @@
-package com.doctorTreat.app.mypage;
+package com.doctorTreat.app.doctorMypage;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
