+++ workspace/src/main/java/com/doctorTreat/app/doctorMypage/DoctorOutController.java
@@ -1,0 +1,5 @@
+package com.doctorTreat.app.doctorMypage;
+
+public class DoctorOutController {
+
+}
