+++ workspace/src/main/java/com/doctorTreat/app/doctorMypage/DoctorUpdateController.java
@@ -1,4 +1,4 @@
-package com.doctorTreat.app.mypage;
+package com.doctorTreat.app.doctorMypage;
 
 public class DoctorUpdateController {
