+++ workspace/src/main/java/com/doctorTreat/app/memberMypage/MemberInfoController.java
@@ -1,0 +1,5 @@
+package com.doctorTreat.app.memberMypage;
+
+public class MemberInfoController {
+
+}
