+++ workspace/src/main/java/com/doctorTreat/app/memberMypage/MemberOutController.java
@@ -1,0 +1,5 @@
+package com.doctorTreat.app.memberMypage;
+
+public class MemberOutController {
+
+}
