+++ workspace/src/main/java/com/doctorTreat/app/memberMypage/MemberPwChangeController.java
@@ -1,4 +1,4 @@
-package com.doctorTreat.app.mypage;
+package com.doctorTreat.app.memberMypage;
 
 public class MemberPwChangeController {
