+++ workspace/src/main/java/com/doctorTreat/app/memberMypage/MemberPwOkController.java
@@ -1,0 +1,5 @@
+package com.doctorTreat.app.memberMypage;
+
+public class MemberPwOkController {
+
+}
