--- workspace/src/main/java/com/doctorTreat/app/mypage/DoctorOutController.java
@@ -1,5 +1,0 @@
-package com.doctorTreat.app.mypage;
-
-public class DoctorOutController {
-
-}
