--- workspace/src/main/java/com/doctorTreat/app/mypage/MemberOutController.java
@@ -1,5 +1,0 @@
-package com.doctorTreat.app.mypage;
-
-public class MemberOutController {
-
-}
